PreFilters were added

+++ src/filters/CompressorAveraging.java
@@ -13,7 +13,7 @@ public class CompressorAveraging extends Compressor {
     @Override
     protected int getData(int index) {
         if (index == 0) return 0;
-        int sum = 0;
+        long sum = 0;
         for (int i = index * compression; i < (index + 1) * compression; i++) {
             if(inputData.get(i) == STAND) {
                 return STAND;
@@ -24,7 +24,7 @@ public class CompressorAveraging extends Compressor {
             if (inputData.get(i) != UNDEFINED)
             sum += inputData.get(i);
         }
-        return sum / compression;
+        return (int)(sum / compression);
     }
