+++ src/filters/CompressorAveragingAbs.java
@@ -13,11 +13,11 @@ public class CompressorAveragingAbs extends Compressor {
     @Override
     protected int getData(int index) {
         if (index == 0) return 0;
-        int sum = 0;
+        long sum = 0;
         for (int i = index * compression; i < (index + 1) * compression; i++) {
             sum += Math.abs(inputData.get(i));
         }
-        return sum / compression;
+        return (int)(sum / compression);
     }
