+++ src/filters/FilterHiPass.java
@@ -7,7 +7,7 @@ import data.DataSet;
  */
 public class FilterHiPass extends FilterBuffered {
     private int indexBefore = -10;
-    private int sumBefore = 0;
+    private long sumBefore = 0;
 
     public FilterHiPass(DataSet inputData, int bufferSize) {
         super(inputData);
@@ -22,7 +22,7 @@ public class FilterHiPass extends FilterBuffered {
         if (index >= size()- bufferSize) {
             return 0;
         }
-        int sum = 0;
+        long sum = 0;
         if(index == (indexBefore +1)) {
            sum = sumBefore + inputData.get(index + bufferSize) - inputData.get(index - bufferSize);
            sumBefore = sum;
@@ -33,7 +33,7 @@ public class FilterHiPass extends FilterBuffered {
                 sum += inputData.get(i);
             }
         }
-        return inputData.get(index) - sum/(2*bufferSize);
+        return inputData.get(index) - (int)(sum/(2*bufferSize));
     }
 
     protected int getData(int index) {
@@ -43,11 +43,11 @@ public class FilterHiPass extends FilterBuffered {
         if (index >= size()-bufferSize) {
             return 0;
         }
-        int sum = 0;
+        long sum = 0;
         for (int i = (index - bufferSize); i < (index + bufferSize); i++) {
             sum += inputData.get(i);
         }
-        return inputData.get(index) - sum/(2*bufferSize);
+        return inputData.get(index) - (int)(sum/(2*bufferSize));
     }
 
 }
